requirements.txt
# ─── Interface ───────────────────────────────────────────────────────────────
streamlit==1.32.0
plotly==5.18.0
pyLDAvis==3.4.1

# ─── PDF Processing ──────────────────────────────────────────────────────────
pymupdf==1.23.26
pdfplumber==0.10.3
pypdf==4.1.0
Pillow==10.2.0

# ─── NLP ─────────────────────────────────────────────────────────────────────
spacy==3.7.4
nltk==3.8.1
gensim==4.3.2
scikit-learn==1.4.1
textstat==0.7.3

# ─── LLM (APIs externas — opcionais) ─────────────────────────────────────────
anthropic==0.21.3
openai==1.14.2

# ─── Dados e utilitários ─────────────────────────────────────────────────────
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.6.3
lxml==5.1.0
tqdm==4.66.2
